refactor: move 5 encoding modules into encoding/
