+++ helix_context/context_manager.py
@@ -32,8 +32,8 @@ from .config import HelixConfig
 from .exceptions import PromoterMismatch
 from .genome import Genome
 from .budget_zone import is_enabled as _budget_zone_is_enabled, zone_cap as _budget_zone_cap
-from .headroom_bridge import compress_text
-from . import legibility
+from .encoding.headroom_bridge import compress_text
+from .encoding import legibility
 from . import session_delivery as _session_delivery
 from .ribosome import DisabledBackend, LiteLLMBackend, Ribosome, OllamaBackend
 from .provenance import apply_metadata_hints, apply_provenance
